requirements.txt
# Core dependencies
PyPDF2>=3.0.0
tiktoken>=0.5.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0

# RAG and LLM components
langchain>=0.0.267
openai>=1.3.0

# Vector DB and embeddings
chromadb>=0.4.18

# Performance tracking
psutil>=5.9.5
tabulate>=0.9.0

# Visualization and analysis
matplotlib>=3.7.2
pandas>=2.0.3

# Web interface (optional)
streamlit>=1.28.0